Z Axis Obstacle Check
added a checkbox like i did to the floor but this one checks along the z axis a certain distance out if anything it in the way, it then prevents the player from snapping to the center axis like normal. Ads the effect of being in front of or behind the 3d objects while in 2d. This will ideally have positive effects on level building in the future.